+++ miqa_worker_task/README.rst
@@ -1,0 +1,1 @@
+Worker task module for MIQA (with learning)
